refactor(course): minor changes
