+++ app/build.gradle.kts
@@ -42,9 +42,4 @@ dependencies {
     implementation("androidx.multidex:multidex:2.0.1")
 }
 
-configurations {
-    all {
-        exclude(group = "com.google.guava", module = "listenablefuture")
-    }
-}
-
+//apply(plugin = "com.google.gms.google-services")
