+++ course/src/main/java/de/hpi/android/course/domain/CourseMapper.kt
