--- course/src/main/java/de/hpi/android/course/presentation/detail/CourseDetailFragment.kt
@@ -1,25 +1,0 @@
-package de.hpi.android.course.presentation.detail
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.lifecycle.ViewModelProviders
-import de.hpi.android.core.presentation.base.BaseFragment
-import de.hpi.android.course.databinding.FragmentCourseDetailBinding
-
-class CourseDetailFragment : BaseFragment<FragmentCourseDetailBinding, CourseDetailViewModel>() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        viewModel = ViewModelProviders.of(this).get(CourseDetailViewModel::class.java)
-    }
-
-    override fun onCreateBinding(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): FragmentCourseDetailBinding {
-        return FragmentCourseDetailBinding.inflate(inflater, container, false).also {
-            it.viewModel = viewModel
-        }
-    }
-}
