+++ course/src/main/java/de/hpi/android/course/presentation/list/CourseListViewModel.kt
