+++ course/src/main/res/layout/fragment_course_list.xml
@@ -2,7 +2,6 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools">
-
     <data>
         <variable
                 name="viewModel"
@@ -26,5 +25,4 @@
                 tools:listitem="@layout/item_course"/>
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
-
 </layout>
